model.ts を分割
