+++ src/app.ts
@@ -1,15 +1,8 @@
 import 'babel-polyfill';
 import Vue from 'vue';
-import App from './components/App.vue';
-import Card from "./components/Card.vue";
-import Player from "./components/Player.vue";
-import Computer from "./components/Computer.vue";
-
-Vue.component('card', Card);
-Vue.component('player', Player);
-Vue.component('computer', Computer);
+import AppComponent from './components/App.vue';
 
 new Vue({
   el: '#app',
-  render: h => h(App)
+  render: h => h(AppComponent)
 });
