+++ src/components/Computer.vue
@@ -281,13 +281,13 @@
 
 <script lang="ts">
 import { Component, Vue, Prop } from "vue-property-decorator";
-import { Computer } from "../models";
+import Computer from "../models/Computer";
 
-@Component
+@Component({ name: "Computer" })
 export default class ComputerComponent extends Vue {
   
   @Prop()
-  public computer!: Computer;
+  public computer: Computer;
 
   public get imagePath() {
     return `images/${this.computer.image}`;
@@ -297,7 +297,6 @@ export default class ComputerComponent extends Vue {
     const match = /#([0-9a-fA-F]{2})([0-9a-fA-F]{2})([0-9a-fA-F]{2})/.exec(
       this.computer.color
     );
-    if (match == null) throw "Invalid color!";
     const colors = [
       parseInt(match[1], 16),
       parseInt(match[2], 16),
